separate the CLI command implementations
