+++ src/CcgVault/CLI/ICliCommand.cs
@@ -1,0 +1,11 @@
+﻿namespace CcgVault
+{
+    partial class Program
+    {
+        interface ICliCommand
+        {
+            void Invoke();
+            void Help(string text);
+        }
+    }
+}
